Rename brave_add_site_dialog_behavior.ts -> brave_add_site_dialog.ts
