+++ browser/resources/settings/brave_add_site_dialog/brave_add_site_dialog.ts
