+++ browser/resources/settings/brave_overrides/add_site_dialog.ts
@@ -3,7 +3,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at https://mozilla.org/MPL/2.0/.
 
-import {BraveAddSiteDialogElement} from '../brave_add_site_dialog/brave_add_site_dialog_behavior.js'
+import {BraveAddSiteDialogElement} from '../brave_add_site_dialog/brave_add_site_dialog.js'
 import {RegisterPolymerComponentReplacement, RegisterPolymerTemplateModifications} from 'chrome://resources/polymer_overriding.js'
 
 RegisterPolymerComponentReplacement(
