+++ browser/resources/settings/sources.gni
@@ -52,7 +52,7 @@ brave_settings_web_component_files = [
 brave_settings_non_web_component_files = [
   "brave_appearance_page/brave_appearance_browser_proxy.ts",
   "brave_clear_browsing_data_dialog/brave_clear_browsing_data_dialog_behavior.ts",
-  "brave_add_site_dialog/brave_add_site_dialog_behavior.ts",
+  "brave_add_site_dialog/brave_add_site_dialog.ts",
   "brave_default_extensions_page/brave_default_extensions_browser_proxy.ts",
   "brave_ipfs_page/brave_ipfs_browser_proxy.ts",
   "brave_new_tab_page/brave_new_tab_browser_proxy.ts",
